+++ Surface_mesh_segmentation/examples/Surface_mesh_segmentation/approx_convex_decomposition_SM_example.cpp
@@ -1,4 +1,4 @@
-#include <CGAL/approx_decomposition.h>
+#include <CGAL/approximate_convex_decomposition.h>
 
 #include <CGAL/Surface_mesh.h>
 #include <CGAL/Exact_predicates_inexact_constructions_kernel.h>
