+++ Surface_mesh_segmentation/examples/Surface_mesh_segmentation/concavity_value_example.cpp
@@ -1,4 +1,4 @@
-#include <CGAL/approx_decomposition.h>
+#include <CGAL/approximate_convex_decomposition.h>
 
 #include <CGAL/Polyhedron_3.h>
 #include <CGAL/Exact_predicates_inexact_constructions_kernel.h>
