+++ Surface_mesh_segmentation/examples/Surface_mesh_segmentation/segmentation_from_sdf_values_concavity_example.cpp
@@ -1,5 +1,5 @@
 #include <CGAL/mesh_segmentation.h>
-#include <CGAL/approx_decomposition.h>
+#include <CGAL/approximate_convex_decomposition.h>
 #include <CGAL/Polyhedron_3.h>
 #include <CGAL/Polyhedron_items_with_id_3.h>
 #include <CGAL/Simple_cartesian.h>
