+++ Surface_mesh_segmentation/test/Surface_mesh_segmentation/approx_decomposition_precomputed_test.cpp
@@ -1,4 +1,4 @@
-#include <CGAL/approx_decomposition.h>
+#include <CGAL/approximate_convex_decomposition.h>
 
 #include <CGAL/Surface_mesh.h>
 #include <CGAL/Exact_predicates_inexact_constructions_kernel.h>
@@ -24,7 +24,7 @@ int main()
     typedef Mesh::Property_map<face_descriptor, int> Clusters_map;
     Clusters_map cluster_ids = mesh.add_property_map<face_descriptor, int>("f:cluster").first;
 
-    std::size_t clusters_num = CGAL::convex_decomposition<Concurrency_tag>(mesh, cluster_ids, 0.3, 1);
+    std::size_t clusters_num = CGAL::approximate_convex_decomposition<Concurrency_tag>(mesh, cluster_ids, 0.3, 1);
 
     std::vector<int> precomputed;
     std::ifstream file("sword_0.3_precomputed.txt");
