+++ Surface_mesh_segmentation/test/Surface_mesh_segmentation/concavity_value_test.cpp
@@ -1,7 +1,7 @@
 #include <CGAL/Polyhedron_3.h>
 #include <CGAL/Exact_predicates_inexact_constructions_kernel.h>
 
-#include <CGAL/approx_decomposition.h>
+#include <CGAL/approximate_convex_decomposition.h>
 
 #include <string>
 #include "Utils.h"
