+++ Surface_mesh_segmentation/test/Surface_mesh_segmentation/concavity_value_test_2.cpp
@@ -1,4 +1,4 @@
-#include <CGAL/approx_decomposition.h>
+#include <CGAL/approximate_convex_decomposition.h>
 
 #include <CGAL/Surface_mesh.h>
 #include <CGAL/Exact_predicates_inexact_constructions_kernel.h>
@@ -23,7 +23,7 @@ int main()
     typedef Mesh::Property_map<face_descriptor, int> Clusters_map;
     Clusters_map cluster_ids = mesh.add_property_map<face_descriptor, int>("f:cluster").first;
 
-    std::size_t clusters_num = CGAL::convex_decomposition<Concurrency_tag>(mesh, cluster_ids, 0.3, 1);
+    std::size_t clusters_num = CGAL::approximate_convex_decomposition<Concurrency_tag>(mesh, cluster_ids, 0.3, 1);
 
     std::cout << "Number of clusters: " << clusters_num << std::endl;
     BOOST_FOREACH(face_descriptor fd, faces(mesh))
